+++ packages/dash-core-components/dash_core_components/Tab.py
@@ -12,7 +12,7 @@ Keyword arguments:
 - id (string; optional): The ID of this component, used to identify dash components
 in callbacks. The ID needs to be unique across all of the
 components in an app.
-- label (optional): The tab's label
+- label (string; optional): The tab's label
 - value (string; optional): Value for determining which Tab is currently selected
 - disabled (boolean; optional): Determines if tab is disabled or not - defaults to false
 - disabled_style (dict; optional): Overrides the default (inline) styles when disabled