+++ packages/dash-core-components/test/test_integration.py
@@ -561,25 +561,25 @@ class Tests(IntegrationTests):
 
         app.layout = html.Div([
             html.H1('Dash Tabs component demo'),
-            dcc.Tabs(id="tabs-example", children=[
-                dcc.Tab(label='Tab One', value='tab-1-example'),
-                dcc.Tab(label='Tab Two', value='tab-2-example'),
+            dcc.Tabs(id="tabs-without-value", children=[
+                dcc.Tab(label='Tab One', value='tab-1'),
+                dcc.Tab(label='Tab Two', value='tab-2'),
             ]),
-            html.Div(id='tabs-content-example')
+            html.Div(id='tabs-content')
         ])
 
 
-        @app.callback(Output('tabs-content-example', 'children'),
-                    [Input('tabs-example', 'value')])
+        @app.callback(Output('tabs-content', 'children'),
+                    [Input('tabs-without-value', 'value')])
         def render_content(tab):
-            if tab == 'tab-1-example':
+            if tab == 'tab-1':
                 return html.H3('Default selected Tab content 1')
-            elif tab == 'tab-2-example':
+            elif tab == 'tab-2':
                 return html.H3('Tab content 2')
 
         self.startServer(app=app)
 
-        default_tab_content = self.wait_for_element_by_css_selector('#tabs-content-example')
+        default_tab_content = self.wait_for_element_by_css_selector('#tabs-content')
 
         self.assertEqual(default_tab_content.text, 'Default selected Tab content 1')